requirements.txt
jsonschema
strictyaml
ruamel.yaml
setuptools
